Adjust emoticon moving speed

+++ EmojiHunter/EmojiHunter/GUIModels/UIEmoticonMoveBehaviors/MeleeMoveBehavior.cs
@@ -14,7 +14,7 @@
 
             if (!HasLeftScreen(uiEmoticon.Position))
             {
-                uiEmoticon.Position += uiEmoticon.Direction * ((int)(DecreaseSpeedFactor * uiEmoticon.GameObject.State.MovementSpeed));
+                uiEmoticon.Position += uiEmoticon.Direction * (DecreaseSpeedFactor * uiEmoticon.GameObject.State.MovementSpeed);
             }
 
             uiEmoticon.Sprite.Position = uiEmoticon.Position;
