+++ EmojiHunter/EmojiHunter/GUIModels/UIEmoticonMoveBehaviors/RunAwayMoveBehavior.cs
@@ -22,7 +22,7 @@
             }
 
             uiEmoticon.Direction = -direction;
-            uiEmoticon.Position += uiEmoticon.Direction * ((int)(DecreaseSpeedFactor * uiEmoticon.GameObject.State.MovementSpeed));
+            uiEmoticon.Position += uiEmoticon.Direction * (DecreaseSpeedFactor * uiEmoticon.GameObject.State.MovementSpeed);
             uiEmoticon.Sprite.Position = uiEmoticon.Position;
         }
     }
